Refactor non-table functions out of Hasura.RQL.DDL.Table
