+++ server/src-exec/Migrate.hs
@@ -4,19 +4,19 @@ module Migrate
   )
 where
 
-import           Data.Time.Clock             (UTCTime)
-import           Language.Haskell.TH.Syntax  (Q, TExp, unTypeQ)
+import           Data.Time.Clock            (UTCTime)
+import           Language.Haskell.TH.Syntax (Q, TExp, unTypeQ)
 
 import           Hasura.Prelude
-import           Hasura.RQL.DDL.Schema.Table
+import           Hasura.RQL.DDL.Schema
 import           Hasura.RQL.Types
 import           Hasura.Server.Query
 
-import qualified Data.Aeson                  as A
-import qualified Data.Text                   as T
-import qualified Data.Yaml.TH                as Y
+import qualified Data.Aeson                 as A
+import qualified Data.Text                  as T
+import qualified Data.Yaml.TH               as Y
 
-import qualified Database.PG.Query           as Q
+import qualified Database.PG.Query          as Q
 
 curCatalogVer :: T.Text
 curCatalogVer = "20"
