+++ server/src-exec/Ops.hs
@@ -10,7 +10,7 @@ import           Migrate                      (curCatalogVer)
 
 import           Hasura.EncJSON
 import           Hasura.Prelude
-import           Hasura.RQL.DDL.Schema.Table
+import           Hasura.RQL.DDL.Schema
 import           Hasura.RQL.Types
 import           Hasura.Server.Query
 import           Hasura.SQL.Types
